+++ src/hrdmc/estimators/__init__.py
@@ -1,4 +1,4 @@
-from hrdmc.estimators.density import estimate_density_profile
+from hrdmc.estimators.density import estimate_density_profile, estimate_open_line_density_profile
 from hrdmc.estimators.local_energy import LocalEnergyResult, estimate_local_energy
 from hrdmc.estimators.pair_distribution import PairDistributionResult, estimate_pair_distribution
 from hrdmc.estimators.structure_factor import StructureFactorResult, estimate_static_structure_factor
@@ -8,6 +8,7 @@ __all__ = [
     "PairDistributionResult",
     "StructureFactorResult",
     "estimate_density_profile",
+    "estimate_open_line_density_profile",
     "estimate_local_energy",
     "estimate_pair_distribution",
     "estimate_static_structure_factor",