+++ src/hrdmc/wavefunctions/__init__.py
@@ -1,3 +1,4 @@
 from hrdmc.wavefunctions.jastrow import HardRodJastrowTrial
+from hrdmc.wavefunctions.trapped import TrappedHardRodTrial
 
-__all__ = ["HardRodJastrowTrial"]
+__all__ = ["HardRodJastrowTrial", "TrappedHardRodTrial"]